+++ tests/MegaphoneComponentTest.php
@@ -30,9 +30,16 @@ it('can render the megaphone component with notification count', function () {
 
     $this->livewire(Megaphone::class)
         ->assertViewIs('megaphone::megaphone')
-        ->assertSeeHtml('<span aria-label="unread count" class="absolute top-0 left-0 aspect-square max-h-fit rounded-full border-2 bg-red-500 px-1.5 shadow leading-5 text-white font-semibold text-xs">
-                1
-            </span>');
+        ->assertSeeHtml('<sub class="absolute top-1 left-1" aria-label="unread count">
+                <span class="relative flex h-5 w-5 -mt-1">
+                    <span class="motion-safe:animate-ping absolute inline-flex h-full w-full rounded-full bg-red-400 opacity-75"></span>
+                    <span class="relative inline-flex rounded-full h-5 w-5 bg-red-500 text-white text-xs aspect-square text-center">
+                        <span class="w-full leading-5">
+                            1
+                        </span>
+                    </span>
+                </span>
+            </sub>');
 });
 
 it('can render the megaphone component with notification dot', function () {
@@ -51,10 +58,12 @@ it('can render the megaphone component with notification dot', function () {
 
     $this->livewire(Megaphone::class)
         ->assertViewIs('megaphone::megaphone')
-        ->assertSeeHtml('<span aria-label="has unread notifications" class="absolute top-0 left-0 aspect-square h-2/5 rounded-full bg-red-500 shadow">')
-        ->assertDontSeeHtml('<span aria-label="unread count" class="absolute top-0 left-0 aspect-square max-h-fit rounded-full border-2 bg-red-500 px-1.5 shadow leading-5 text-white font-semibold text-xs">
-                1
-            </span>');
+        ->assertSeeHtml('<sub class="absolute top-2 left-2" aria-label="has unread notifications">
+                <span class="relative flex h-3 w-3 -mt-1">
+                  <span class="motion-safe:animate-ping absolute inline-flex h-full w-full rounded-full bg-red-400 opacity-75"></span>
+                  <span class="relative inline-flex rounded-full h-3 w-3 bg-red-500"></span>
+                </span>
+            </sub>');
 });
 
 it('can load announcements', function () {